refactor: Decompose patients page into components
