+++ .env
@@ -1,1 +1,5 @@
-NODE_ENV=development
+AUTH0_SECRET=1f69d8cbb4f1eab8b4dea6ac2dcb80a1c684481a8af7317bc4c64a11256de030
+AUTH0_BASE_URL='http://localhost:3000'
+AUTH0_ISSUER_BASE_URL='https://sabai.jp.auth0.com'
+AUTH0_CLIENT_ID='jNlvcpCDrfAChWX2nsEp9kCk0WjRFmSt'
+AUTH0_CLIENT_SECRET='Z7U0t0vyliuUTbcrRE2OGGsM51jlKkvi8-zZHCqqJGLfFq1X0vMNIF8UgRl6IwgQ'
