+++ pages/api/auth/[auth0].js
@@ -1,3 +1,3 @@
-import { handleAuth } from '@auth0/nextjs-auth0';
+import { handleAuth } from "@auth0/nextjs-auth0";
 
 export default handleAuth();
