+++ utils/constants.js
@@ -1,3 +1,10 @@
-export const API_URL = "https://projectsabai-vza8.onrender.com";
+// export const API_URL = "https://projectsabai-vza8.onrender.com";
 export const CLOUDINARY_URL = "https://res.cloudinary.com/dxy0byphl";
-//export const API_URL = "http://localhost:8000";
+export const API_URL = "http://localhost:8000";
+
+export const venueOptions = {
+  SV: "Smong Village",
+  SO: "Smong Orphanage",
+  SPS: "Smong Primary School",
+  SSS: "Smong Secondary School",
+};
